requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow-gpu>=1.3.0
keras>=2.0.8
opencv-python
h5py
IPython[all]
pika
pandas
tqdm